services/user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "User Servic for SocialX"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "django-simple-history>=3.10.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-crontab>=3.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# Include only the desired packages, potentially excluding 'static'
include = ["coreAPI", "coreAPI.*"]
# You can also use 'exclude'
exclude = ["static", "tests*"]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.7",
]
